requirements.txt
torch==2.8.0
torchvision==0.23.0
pandas>=1.5.0,<2.3.0
numpy>=1.24,<2.0
pillow>=9.5
scikit-learn>=1.2,<1.6
albumentations==1.4.15
albucore==0.0.15
opencv-python-headless>=4.7,<5.0
scikit-image>=0.22,<0.25
matplotlib>=3.6,<3.9
fpdf2>=2.7.0
segmentation-models-pytorch==0.3.3
timm==0.9.2
pretrainedmodels==0.7.4
six
tqdm